Change from SubTasks to SubTaskIds

+++ TaskManagementSystem.Server/Migrations/001_create_task_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS "task" (
   "name" text not null,
   "description" text,
   "status" int not null,
-  "sub_tasks" int[] not null,
+  "sub_tasks" bigint[] not null,
   "created_at" timestamp not null,
   "updated_at" timestamp not null
 );
